fix: Fix compiler warnings in release reducers JS
